api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy>=1.24.0
xgboost==2.0.3
scikit-learn==1.3.2
rdkit-pypi==2023.3.1b1
pennylane==0.32.0
openai==1.3.7
pandas==2.1.3
Pillow==10.1.0

